refatoraçao do grafico
