+++ frontend/src/components/AreaChart/styled.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const Div = styled.div`
+    display: flex;
+    align-items: center;
+
+    #title-dashboard {
+        margin-right: 20px;
+    }
+`
+export const ContainerChart = styled.div`
+    margin: 0 auto;
+`;
