+++ frontend/src/pages/DataAnalysis/styled.tsx
@@ -9,11 +9,8 @@ export const Container = styled.div`
   height: 100vh;
   display: flex;
   flex-direction: column;
-  align-items: center;
-  justify-content: center;
+  text-align: center;
   color: var(--white);
-
-  margin-top: 50px;
 `;
 
 export const Div = styled.div``;
